+++ NEWS.md
@@ -1,3 +1,9 @@
+# autometric 0.1.0
+
+* Adjust default `ylim` in `log_plot()`.
+* Add `log_phase_get()`, `log_phase_set()`, and `log_phase_reset()` (#2, #3, @koefoeden).
+* Allow some log files to be empty in `log_read()`.
+
 # autometric 0.0.5
 
 * Check the validity of the file path in `log_print()` and `log_start()` (#1, @shikokuchuo).